Backend/requirements.txt
watchdog
psutil
schedule
streamlit
boto3
pandas
python-dotenv
